Refactor: Modularize components and optimize project structure
This commit introduces a comprehensive restructuring of the project to improve modularity, maintainability, and scalability. Key changes include:

1. Separation of concerns into dedicated modules:
   - Created main.py as the application entry point responsible for initialization and settings
   - Extracted video conversion functionality into video_converter.py
   - Established separate modules for audio and image processing (audio_converter.py, image_converter.py)
   - Integrated language settings into the configuration module

2. Optimized module dependencies:
   - Each component now depends only on necessary interfaces
   - Centralized coordination logic in the main application

3. Enhanced configuration management:
   - Implemented persistent storage for language preferences
   - Extracted all UI text into language dictionaries
   - Added support for dynamic language switching with real-time UI updates

4. Resolved existing issues:
   - Fixed tab index out-of-bounds error
   - Improved configuration file loading/saving logic

These changes result in a more modular architecture that is easier to maintain and extend. The codebase is now better organized, with clear separation between presentation, business logic, and configuration.

Related Issue: #123 (Project Structure Refactoring)
