+++ config.json
@@ -1,0 +1,1 @@
+{"language": "简体中文"}
