+++ locales/zh_CN.json
@@ -1,0 +1,47 @@
+{
+    "input_folder_label": "输入文件夹:",
+    "output_folder_label": "输出文件夹:",
+    "target_format_label": "目标格式:",
+    "include_subfolders": "包含子文件夹",
+    "overwrite_existing": "覆盖已存在文件",
+    "thread_count_label": "并发线程数:",
+    "start_conversion": "开始转换",
+    "status_label": "状态:",
+    "progress_label": "进度:",
+    "file_stats_label": "文件统计:",
+    "log_label": "转换日志:",
+    "browse_button": "浏览...",
+    "auto_set_button": "自动设置",
+    "error_no_input_folder": "错误",
+    "error_no_input_folder_msg": "请选择输入文件夹",
+    "error_invalid_input_folder": "错误",
+    "error_invalid_input_folder_msg": "输入文件夹不存在: ",
+    "error_no_output_folder": "错误",
+    "error_no_output_folder_msg": "请选择输出文件夹",
+    "error_create_output_folder": "错误",
+    "error_create_output_folder_msg": "无法创建输出文件夹: ",
+    "error_ffmpeg": "错误",
+    "error_ffmpeg_msg": "无法执行FFmpeg: ",
+    "no_video_files": "未找到视频文件",
+    "no_video_files_msg": "未找到视频文件在目录: ",
+    "conversion_start": "正在查找视频文件...",
+    "conversion_start_log": "开始扫描视频文件...",
+    "video_files_found": "找到 {} 个视频文件",
+    "using_threads": "使用 {} 个线程进行转换",
+    "target_format": "目标格式: {}",
+    "output_directory": "输出目录: {}",
+    "converting": "正在转换...",
+    "conversion_complete": "转换完成",
+    "conversion_complete_msg": "转换完成！共处理 {} 个文件，成功 {} 个，失败 {} 个。",
+    "conversion_failed": "转换失败",
+    "conversion_failed_msg": "转换过程中发生错误: ",
+    "file_conversion_start": "开始转换: {}",
+    "file_conversion_success": "成功: {}",
+    "file_conversion_error": "错误: 处理 {} 时出错 - {}",
+    "file_conversion_skip": "跳过: {} 已存在",
+    "file_not_found": "错误: {} 不是有效的文件",
+    "unknown_error": "未知错误: 转换 {} 时发生异常 - {}",
+    "conversion_finished_title": "完成",
+    "conversion_finished_msg": "转换完成！\n成功: {}\n失败: {}",
+    "set_output_folder": "自动设置输出文件夹为: {}"
+}
